Allow to the player to open the gates
The player can now open the gates via their security terminal. There is one terminal by gate. The next work is to animate the opening and the closing. Nota bene : the fence are a little glitchy when a gate is open then close, we can pass through certain tiles... I'll gonna work in the future on this bug.