Gates animation - Update to obtain the best render
The render now respect what we thought the last time. The gates are opening from the middle to the sides and are closing from the sides to the middle. The last feature is to check if there are some stuff on the gate passage : if there are, so the gate can't close.